Fix:vehicle_gpsd:Fix warning.
git-svn-id: http://svn.code.sf.net/p/navit/code/trunk/navit@6011 ffa7fe5e-494d-0410-b361-a75ebd5db220

+++ navit/vehicle/gpsd/vehicle_gpsd.c
@@ -101,7 +101,7 @@ vehicle_gpsd_callback(struct gps_data_t *data, const char *buf, size_t len,
 			}
 		}
 	}	
-	dbg(lvl_debug,"data->set="LONGLONG_HEX_FMT"\n", data->set);
+	dbg(lvl_debug,"data->set="LONGLONG_HEX_FMT"\n", (unsigned long long)data->set);
 	if (data->set & SPEED_SET) {
 		priv->speed = data->fix.speed * 3.6;
 		if(!isnan(data->fix.speed))
